Took out of the component
