+++ src/StyledComponents/Global.ts
@@ -1,0 +1,32 @@
+import { createGlobalStyle } from "styled-components";
+
+const Global = createGlobalStyle`
+  * {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+
+  a {
+    text-decoration: none;
+    color: unset;
+  }
+
+  ::-webkit-scrollbar {
+    width: 5px;
+  }
+
+  ::-webkit-scrollbar-track {
+    background: white;
+  }
+
+  ::-webkit-scrollbar-thumb {
+    background: #B5B5B5;
+  }
+
+  ::-webkit-scrollbar-thumb:hover {
+    background: #808080;
+  }
+`;
+
+export default Global;
